move query components to unified folder
